# contracts/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "Voice Platform API",
    "version": "1.0.0-draft",
    "description": "Versioned foundation contract. Feature endpoints are added only with matching ACL and contract tests."
  },
  "paths": {
...
    "/api/v1/channels/{channelID}/messages": {
      "get": {
        "operationId": "listTextMessages",
        "description": "Authenticated member only. Returns newest-first cursor pages from a current TEXT channel. Deleted messages retain an ID and deleted marker but never expose their former body.",
        "parameters": [
          { "name": "channelID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "before", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Cursor page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TextMessagePage" } } } },
          "400": { "description": "Malformed page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Unavailable text channel", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "operationId": "createTextMessage",
        "description": "Authenticated member only. Creates or returns the same message for an author-and-channel-scoped client_message_id. The current channel must be non-archived TEXT; a reply may reference only that same channel.",
        "parameters": [
          { "name": "channelID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TextMessageCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Message created or idempotently returned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TextMessage" } } } },
          "400": { "description": "Malformed or oversized message", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Unavailable text channel or cross-channel reply", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/admin/categories/order": {
      "put": {
        "operationId": "reorderCategories",
        "description": "Administrator-only. The IDs must be the complete category order for the supplied topology revision.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryReorderRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Order changed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TopologyRevision" } } } },
          "400": { "description": "Malformed or duplicate order entries", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No valid active session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Administrator role required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Revision or topology mismatch; refresh state and retry", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/admin/channels/{channelID}/category": {
      "patch": {
        "operationId": "moveChannel",
        "description": "Administrator-only. Moves the existing channel without changing its TEXT or VOICE kind.",
        "parameters": [
          { "name": "channelID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ChannelMoveRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Channel moved to the end of target category", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChannelMoveResult" } } } },
          "400": { "description": "Malformed move request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No valid active session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Administrator role required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Revision or topology mismatch; refresh state and retry", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/admin/categories/{categoryID}/channels/order": {
      "put": {
        "operationId": "reorderChannels",
        "description": "Administrator-only. IDs must be every channel currently in this one category for the supplied topology revision.",
        "parameters": [
          { "name": "categoryID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChannelReorderRequest" } } }
        },
        "responses": {
          "200": { "description": "Channel order changed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TopologyRevision" } } } },
          "400": { "description": "Malformed or duplicate order entries", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No valid active session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Administrator role required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Revision, category or channel membership mismatch; refresh state and retry", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TextMessageCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["client_message_id", "body"],
        "properties": {
          "client_message_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string", "minLength": 1, "maxLength": 8000 },
          "reply_to_id": { "type": "string", "format": "uuid" },
          "attachment_ids": { "type": "array", "maxItems": 10, "uniqueItems": true, "items": { "type": "string", "format": "uuid" } }
        }
      },
      "TextMessage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "channel_id", "author_id", "client_message_id", "body", "revision", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "channel_id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "client_message_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string" },
          "revision": { "type": "integer", "minimum": 1 },
          "reply_to_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "TextMessagePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["messages"],
        "properties": {
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/TextMessageHistoryItem" } },
          "next_cursor": { "type": "string", "format": "uuid" }
        }
      },
...
      "TextMessageHistoryItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "channel_id", "author_id", "client_message_id", "body", "created_at", "revision", "deleted", "attachments"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "channel_id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "client_message_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string" },
          "reply_to_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" },
          "edited_at": { "type": "string", "format": "date-time" },
          "revision": { "type": "integer", "minimum": 1 },
          "deleted": { "type": "boolean" },
          "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/TextMessageAttachment" } }
        }
      },
      "TextMessageAttachment": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "original_name", "byte_size"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "original_name": { "type": "string" },
          "byte_size": { "type": "integer", "minimum": 0, "maximum": 25000000 }
        }
      },
...
      "CategoryReorderRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_revision", "ids"],
        "properties": {
          "expected_revision": { "type": "integer", "minimum": 1 },
          "ids": { "type": "array", "items": { "type": "string", "format": "uuid" }, "uniqueItems": true }
        }
      },
      "TopologyRevision": {
        "type": "object",
        "additionalProperties": false,
        "required": ["revision"],
        "properties": {
          "revision": { "type": "integer", "minimum": 1 }
        }
      },
...
      "ChannelMoveRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["category_id", "expected_revision"],
        "properties": {
          "category_id": { "type": "string", "format": "uuid" },
          "expected_revision": { "type": "integer", "minimum": 1 }
        }
      },
      "ChannelMoveResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "category_id", "position", "revision"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "category_id": { "type": "string", "format": "uuid" },
          "position": { "type": "integer", "minimum": 0 },
          "revision": { "type": "integer", "minimum": 1 }
        }
      },
      "ChannelReorderRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_revision", "ids"],
        "properties": {
          "expected_revision": { "type": "integer", "minimum": 1 },
          "ids": { "type": "array", "items": { "type": "string", "format": "uuid" }, "uniqueItems": true }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": {
            "type": "string",
            "enum": ["UNAUTHENTICATED", "ACCOUNT_BLOCKED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "VALIDATION_FAILED", "RATE_LIMITED", "MAINTENANCE", "MAINTENANCE_UNAVAILABLE", "ATTACHMENT_TOO_LARGE", "INSUFFICIENT_STORAGE", "ACTIVE_VOICE_LEASE", "VOICE_CHANNEL_UNAVAILABLE", "VOICE_LEASE_UNAVAILABLE", "INTERNAL"]
          },
          "message": { "type": "string" },
          "request_id": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      }
    }
  }
}